Tasks completed and alt img info added
Update the logo path to display the rock climbing logo in the images
folder.
Link the logo on all four pages to connect to the home page.
On the home page, add the images paths for the three images in the
climbing subdirectory:
Man on Wall
Climbing Gear
Chalked Hands
On the home page, link each of the three images to their appropriate
pages. The link should open the page into the existing window.
Man on Wall -> man.html
Climbing Gear -> gear.html
Chalked Hands -> chalk.html
On the home page, link the key words in the paragraphs to the following
web sites:

and alt img info added